# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feature-requests/{id}:
    get:
      operationId: getFeatureRequest
      tags: [feature-requests]
      summary: Get a single feature request (owner or admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feature request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureRequest"
        "404":
          description: Not found
    patch:
      operationId: updateFeatureRequest
      tags: [feature-requests]
      summary: Update a feature request (status, priority)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureRequestUpdate"
      responses:
        "200":
          description: Updated feature request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureRequest"
        "404":
          description: Not found
    delete:
      operationId: deleteFeatureRequest
      tags: [feature-requests]
      summary: Delete a feature request and its underlying conversation (owner or admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /feature-requests/{id}/versions:
    get:
      operationId: listFeatureRequestVersions
      tags: [feature-requests]
      summary: List the version history of a feature request (owner or admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Versions, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureRequestVersion"
        "404":
          description: Not found
  /feature-requests/{id}/resynthesize:
    post:
      operationId: resynthesizeFeatureRequest
      tags: [feature-requests]
      summary: Re-run the AI to update the requirements doc with the latest conversation context
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated feature request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureRequest"
        "404":
          description: Not found
...
components:
  schemas:
...
    FeatureRequest:
      type: object
      required:
        - id
        - userId
        - userEmail
        - title
        - summary
        - problem
        - benefits
        - currentSpend
        - scope
        - status
        - priority
        - createdAt
        - updatedAt
        - conversationId
        - minor
      properties:
        id:
          type: integer
        userId:
          type: string
        userEmail:
          type: string
        userName:
          type: ["string", "null"]
        title:
          type: string
        summary:
          type: string
        problem:
          type: string
          description: Why this is a problem for the customer
        benefits:
          type: string
          description: Benefits/value the feature delivers
        currentSpend:
          type: string
          description: Money/time the customer spends today on this problem
        scope:
          type: string
          description: Full scoped requirements as Markdown
        status:
          type: string
          enum: [requested, ready_for_execution, planned, in_progress, deployed]
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        conversationId:
          type: integer
        minor:
          type: boolean
          description: Filed via the "Minor" quick-report path (small bug/feature).
        aiPriorityRank:
          type: ["integer", "null"]
          description: 1-based rank assigned by the periodic AI prioritization job (only set for "requested" items in the most recent run).
        aiPriorityRationale:
          type: ["string", "null"]
          description: Short explanation of why this request received its AI rank.
        aiPriorityGeneratedAt:
          type: ["string", "null"]
          format: date-time
          description: When the current AI rank was generated.
        relatedRequestIds:
          type: ["array", "null"]
          description: Admin-only. Ids of other feature requests this looks like a near-duplicate of or strongly clusters with. Surface only in admin UIs.
          items:
            type: integer
        clusterRationale:
          type: ["string", "null"]
          description: Admin-only. Short explanation of how this request relates to the items in relatedRequestIds. Never shown to the requesting customer.
        adminPriorityRank:
          type: ["integer", "null"]
          description: Admin-only. Manual 1-based rank set by an admin when moving a request to "Planned". Lower = higher priority. Used to order the backlog fed to the external dev agent. Defaults to the AI rank but can be overridden.
        paperclipIssueId:
          type: ["string", "null"]
          description: Admin-only. ID of the Paperclip issue created for this request, or null if not yet pushed.
        paperclipIdentifier:
          type: ["string", "null"]
          description: Admin-only. Human-readable Paperclip identifier (e.g. "PUL-23").
        paperclipStatus:
          type: ["string", "null"]
          description: Admin-only. Latest Paperclip status (open/in_progress/in_review/done/cancelled/...).
        paperclipPushedAt:
          type: ["string", "null"]
          format: date-time
          description: Admin-only. When this request was first pushed to Paperclip.
        paperclipLastSyncedAt:
          type: ["string", "null"]
          format: date-time
          description: Admin-only. When the Paperclip snapshot was last refreshed by the poller.
        paperclipPushError:
          type: ["string", "null"]
          description: Admin-only. Last push error message, if the most recent push attempt failed.
        engineeringOwner:
          type: ["string", "null"]
          enum: [agent, human, null]
          description: Admin-only. Engineering destination chosen at plan-time. "agent" routes to Paperclip + the dev-agent queue; "human" routes to Notion only. null = not yet routed (legacy rows are treated as agent-eligible).
        notionPageId:
          type: ["string", "null"]
          description: Admin-only. ID of the Notion page created for this human-routed request, or null if not yet pushed.
        notionUrl:
          type: ["string", "null"]
          description: Admin-only. Direct URL to the Notion page.
        notionStatus:
          type: ["string", "null"]
          description: Admin-only. Latest Notion Status value (e.g. Unpicked / In progress / Done).
        notionAssignee:
          type: ["string", "null"]
          description: Admin-only. Display name of the human assignee synced from Notion, if any.
        notionPushedAt:
          type: ["string", "null"]
          format: date-time
          description: Admin-only. When this request was first pushed to Notion.
        notionLastSyncedAt:
          type: ["string", "null"]
          format: date-time
          description: Admin-only. When the Notion snapshot was last refreshed by the poller.
        notionPushError:
          type: ["string", "null"]
          description: Admin-only. Last Notion push error message, if the most recent push attempt failed.
    FeatureRequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [requested, ready_for_execution, planned, in_progress, deployed]
        priority:
          type: string
          enum: [low, medium, high]
        adminPriorityRank:
          type: ["integer", "null"]
          description: Admin-set priority rank for the Planned backlog. Pass null to clear.
          minimum: 1
        engineeringOwner:
          type: ["string", "null"]
          enum: [agent, human, null]
          description: Admin-only. Engineering destination set when moving a request to "Planned". "agent" → Paperclip + dev-agent queue; "human" → Notion only.
    FeatureRequestVersion:
      type: object
      required:
        - id
        - featureRequestId
        - versionNumber
        - title
        - summary
        - problem
        - benefits
        - currentSpend
        - scope
        - priority
        - changeReason
        - createdAt
      properties:
        id:
          type: integer
        featureRequestId:
          type: integer
        versionNumber:
          type: integer
        title:
          type: string
        summary:
          type: string
        problem:
          type: string
        benefits:
          type: string
        currentSpend:
          type: string
        scope:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        changeReason:
          type: string
        createdByUserId:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...